Removed unnecessary casts.

+++ vassal-app/src/main/java/VASSAL/tools/hex/OffsetCoord.java
@@ -17,7 +17,7 @@ public class OffsetCoord {
 
   public static OffsetCoord qoffsetFromCube(int offset, Hex h) {
     final int col = h.q;
-    final int row = h.r + (int) ((h.q + offset * (h.q & 1)) / 2);
+    final int row = h.r + ((h.q + offset * (h.q & 1)) / 2);
     if (offset != OffsetCoord.EVEN && offset != OffsetCoord.ODD) {
       throw new IllegalArgumentException("offset must be EVEN (+1) or ODD (-1)");
     }
@@ -26,7 +26,7 @@ public class OffsetCoord {
 
   public static Hex qoffsetToCube(int offset, OffsetCoord h) {
     final int q = h.col;
-    final int r = h.row - (int) ((h.col + offset * (h.col & 1)) / 2);
+    final int r = h.row - ((h.col + offset * (h.col & 1)) / 2);
     final int s = -q - r;
     if (offset != OffsetCoord.EVEN && offset != OffsetCoord.ODD) {
       throw new IllegalArgumentException("offset must be EVEN (+1) or ODD (-1)");
@@ -35,7 +35,7 @@ public class OffsetCoord {
   }
 
   public static OffsetCoord roffsetFromCube(int offset, Hex h) {
-    final int col = h.q + (int) ((h.r + offset * (h.r & 1)) / 2);
+    final int col = h.q + ((h.r + offset * (h.r & 1)) / 2);
     final int row = h.r;
     if (offset != OffsetCoord.EVEN && offset != OffsetCoord.ODD) {
       throw new IllegalArgumentException("offset must be EVEN (+1) or ODD (-1)");
@@ -44,7 +44,7 @@ public class OffsetCoord {
   }
 
   public static Hex roffsetToCube(int offset, OffsetCoord h) {
-    final int q = h.col - (int) ((h.row + offset * (h.row & 1)) / 2);
+    final int q = h.col - ((h.row + offset * (h.row & 1)) / 2);
     final int r = h.row;
     final int s = -q - r;
     if (offset != OffsetCoord.EVEN && offset != OffsetCoord.ODD) {
